requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.4
pydantic-settings>=2.3.4
python-dotenv==1.0.0
Pillow>=10.2.0
requests==2.31.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0